### What does this PR do?
Adds the ability to create replica 3 arbiter 1 gluster volumes

### What issues does this PR fix or reference?
 #42955

### Previous Behavior
Arbiter configuration was not exposed in glusterfs volume creation

### New Behavior
Volumes can be created with arbiter (metadata-only replication) bricks

### Tests written?

No

### Commits signed with GPG?

No

Please review [Salt's Contributing Guide](https://docs.saltstack.com/en/latest/topics/development/contributing.html) for best practices.

See GitHub's [page on GPG signing](https://help.github.com/articles/signing-commits-using-gpg/) for more information about signing commits with GPG.


```
jdito@sum-dev-03:/data/salt/salt/gluster$ cat cluster-startup.sls 
peer-cluster:
  glusterfs.peered:
    - names:
      - sum-glfiler-01.storage.dy.gl
      - sum-glfiler-02.storage.dy.gl
      - sum-glfiler-03.storage.dy.gl

{% for volume in ['users','public'] %}
{{ volume }}:
  glusterfs.volume_present:
    - bricks:
      - sum-glfiler-01.storage.dy.gl:/gluster/{{ volume }}/brick
      - sum-glfiler-02.storage.dy.gl:/gluster/{{ volume }}/brick
      - sum-glfiler-03.storage.dy.gl:/gluster/{{ volume }}/brick
    - replica: 3
    - arbiter: True
    - start: True
    - force: True

{% endfor %}
```
```
[root@sum-glfiler-03 ~]# salt-call state.apply gluster.cluster-startup
local:
----------
          ID: peer-cluster
    Function: glusterfs.peered
        Name: sum-glfiler-01.storage.dy.gl
      Result: True
     Comment: Host sum-glfiler-01.storage.dy.gl already peered
     Started: 15:31:24.035879
    Duration: 64.799 ms
     Changes:   
----------
          ID: peer-cluster
    Function: glusterfs.peered
        Name: sum-glfiler-02.storage.dy.gl
      Result: True
     Comment: Host sum-glfiler-02.storage.dy.gl already peered
     Started: 15:31:24.101094
    Duration: 58.168 ms
     Changes:   
----------
          ID: peer-cluster
    Function: glusterfs.peered
        Name: sum-glfiler-03.storage.dy.gl
      Result: True
     Comment: Peering with localhost is not needed
     Started: 15:31:24.159656
    Duration: 33.415 ms
     Changes:   
----------
          ID: users
    Function: glusterfs.volume_present
      Result: True
     Comment: Volume users is created and is started
     Started: 15:31:24.193509
    Duration: 4356.89 ms
     Changes:   
              ----------
              new:
                  - users
              old:
----------
          ID: public
    Function: glusterfs.volume_present
      Result: True
     Comment: Volume public is created and is started
     Started: 15:31:28.550847
    Duration: 3752.007 ms
     Changes:   
              ----------
              new:
                  - public
                  - users
              old:
                  - users

Summary for local
------------
Succeeded: 5 (changed=2)
Failed:    0
------------
Total states run:     5
Total run time:   8.265 s
[root@sum-glfiler-03 ~]# gluster volume info
 
Volume Name: public
Type: Replicate
Volume ID: 395e9b53-434f-4a07-9f55-5500ba7499ff
Status: Started
Snapshot Count: 0
Number of Bricks: 1 x (2 + 1) = 3
Transport-type: tcp
Bricks:
Brick1: sum-glfiler-01.storage.dy.gl:/gluster/public/brick
Brick2: sum-glfiler-02.storage.dy.gl:/gluster/public/brick
Brick3: sum-glfiler-03.storage.dy.gl:/gluster/public/brick (arbiter)
Options Reconfigured:
transport.address-family: inet
nfs.disable: on
performance.client-io-threads: off
 
Volume Name: users
Type: Replicate
Volume ID: 787af828-f09b-47fb-b7b9-88486899438d
Status: Started
Snapshot Count: 0
Number of Bricks: 1 x (2 + 1) = 3
Transport-type: tcp
Bricks:
Brick1: sum-glfiler-01.storage.dy.gl:/gluster/users/brick
Brick2: sum-glfiler-02.storage.dy.gl:/gluster/users/brick
Brick3: sum-glfiler-03.storage.dy.gl:/gluster/users/brick (arbiter)
Options Reconfigured:
transport.address-family: inet
nfs.disable: on
performance.client-io-threads: off
```